requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-google-genai>=0.1.0
supabase>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
duckduckgo-search>=5.0.0
ddgs>=1.0.0
arxiv>=2.1.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
deep-translator>=1.8.0
pydantic>=2.0.0
markdown>=3.0.0
python-docx>=1.0.0
pypdf>=3.0.0
